+++ google/cloud/compute/backend_services/v1/internal/backend_services_rest_connection_impl.h
@@ -91,6 +91,10 @@ class BackendServicesRestConnectionImpl
   ListBackendServices(google::cloud::cpp::compute::backend_services::v1::
                           ListBackendServicesRequest request) override;
 
+  StreamRange<google::cloud::cpp::compute::v1::BackendService> ListUsable(
+      google::cloud::cpp::compute::backend_services::v1::ListUsableRequest
+          request) override;
+
   future<StatusOr<google::cloud::cpp::compute::v1::Operation>>
   PatchBackendService(google::cloud::cpp::compute::backend_services::v1::
                           PatchBackendServiceRequest const& request) override;