+++ google/cloud/compute/backend_services/v1/internal/backend_services_rest_metadata_decorator.cc
@@ -127,6 +127,15 @@ BackendServicesRestMetadata::ListBackendServices(
   return child_->ListBackendServices(rest_context, request);
 }
 
+StatusOr<google::cloud::cpp::compute::v1::BackendServiceListUsable>
+BackendServicesRestMetadata::ListUsable(
+    rest_internal::RestContext& rest_context,
+    google::cloud::cpp::compute::backend_services::v1::ListUsableRequest const&
+        request) {
+  SetMetadata(rest_context);
+  return child_->ListUsable(rest_context, request);
+}
+
 future<StatusOr<google::cloud::cpp::compute::v1::Operation>>
 BackendServicesRestMetadata::AsyncPatchBackendService(
     CompletionQueue& cq,