+++ google/cloud/compute/backend_services/v1/internal/backend_services_rest_stub.h
@@ -93,6 +93,11 @@ class BackendServicesRestStub {
                       google::cloud::cpp::compute::backend_services::v1::
                           ListBackendServicesRequest const& request) = 0;
 
+  virtual StatusOr<google::cloud::cpp::compute::v1::BackendServiceListUsable>
+  ListUsable(google::cloud::rest_internal::RestContext& rest_context,
+             google::cloud::cpp::compute::backend_services::v1::
+                 ListUsableRequest const& request) = 0;
+
   virtual future<StatusOr<google::cloud::cpp::compute::v1::Operation>>
   AsyncPatchBackendService(
       google::cloud::CompletionQueue& cq,
@@ -203,6 +208,11 @@ class DefaultBackendServicesRestStub : public BackendServicesRestStub {
   ListBackendServices(google::cloud::rest_internal::RestContext& rest_context,
                       google::cloud::cpp::compute::backend_services::v1::
                           ListBackendServicesRequest const& request) override;
+
+  StatusOr<google::cloud::cpp::compute::v1::BackendServiceListUsable>
+  ListUsable(google::cloud::rest_internal::RestContext& rest_context,
+             google::cloud::cpp::compute::backend_services::v1::
+                 ListUsableRequest const& request) override;
 
   future<StatusOr<google::cloud::cpp::compute::v1::Operation>>
   AsyncPatchBackendService(