+++ google/cloud/compute/backend_services/v1/internal/backend_services_tracing_connection.h
@@ -80,6 +80,10 @@ class BackendServicesTracingConnection
   ListBackendServices(google::cloud::cpp::compute::backend_services::v1::
                           ListBackendServicesRequest request) override;
 
+  StreamRange<google::cloud::cpp::compute::v1::BackendService> ListUsable(
+      google::cloud::cpp::compute::backend_services::v1::ListUsableRequest
+          request) override;
+
   future<StatusOr<google::cloud::cpp::compute::v1::Operation>>
   PatchBackendService(google::cloud::cpp::compute::backend_services::v1::
                           PatchBackendServiceRequest const& request) override;