+++ google/cloud/compute/security_policies/v1/security_policies_client.cc
@@ -204,6 +204,7 @@ SecurityPoliciesClient::ListPreconfiguredExpressionSets(
 future<StatusOr<google::cloud::cpp::compute::v1::Operation>>
 SecurityPoliciesClient::PatchSecurityPolicy(
     std::string const& project, std::string const& security_policy,
+    std::string const& update_mask,
     google::cloud::cpp::compute::v1::SecurityPolicy const&
         security_policy_resource,
     Options opts) {
@@ -212,6 +213,7 @@ SecurityPoliciesClient::PatchSecurityPolicy(
       request;
   request.set_project(project);
   request.set_security_policy(security_policy);
+  request.set_update_mask(update_mask);
   *request.mutable_security_policy_resource() = security_policy_resource;
   return connection_->PatchSecurityPolicy(request);
 }